Reorganizing routes.
