+++ package.json
@@ -12,6 +12,7 @@
     "dotenv": "^4.0.0",
     "express": "~4.15.2",
     "express-session": "^1.15.6",
+    "mongoose": "^4.12.3",
     "morgan": "~1.8.1",
     "nodemon": "^1.12.1",
     "pug": "~2.0.0-beta11",
