+++ routes/index.js
@@ -1,54 +1,21 @@
 require('dotenv').config();
 var express = require('express');
 let request = require('request-promise');
-
-const { fetchUpcomingParashot } = require('../lib/hebcal');
+const secure = require('./secure');
 
 var router = express.Router();
 
 const checkAuth = (req, res, next) => {
   if (!req.session.authorization) {
-    res.redirect('/connect');
+    res.redirect('/');
   } else {
     next();
   }
 }
 
-/* GET home page. */
-router.get('/', checkAuth, (req, res, next) => {
-  const options = {
-    uri: 'https://oauth.reddit.com/api/v1/me',
-    method: 'GET',
-    headers: {
-      'Authorization': `bearer ${req.session.authorization.access_token}`,
-      'User-Agent': 'webapp for the /r/judaism dvar torah project by /r/o_m_f_g'
-    },
-    json: true,
-  };
+router.use('/secure', checkAuth, secure);
 
-  Promise.all([
-    request(options),
-    fetchUpcomingParashot()
-  ])
-  .then(([me, parashot]) => {
-      res.render('home', {
-      title: 'register to write a dvar.',
-      name: me.name,
-      parashot,
-    })
-  })
-  .catch((error) => {
-    console.error(error);
-  });
-});
-
-router.post('/submit', (req, res, next) => {
-  const { parashah, name } = req.body;
-
-  // insert into database.
-});
-
-router.get('/connect', (req, res, next) => {
+router.get('/', (req, res, next) => {
   res.render('index', { title: 'connect to reddit.'})
 });
 
@@ -72,7 +39,7 @@ router.get('/authorize', (req, res, next) => {
     })
     .then((response) => {
       req.session.authorization = response;
-      res.redirect('/');
+      res.redirect('/secure');
     })
     .catch((error) => {
       console.log(error);
